index.tsをserver.tsとtransportsディレクトリに分割
- src/server.ts: MCPサーバーファクトリ(ツール定義・ハンドラ・エラー整形)
- src/transports/stdio.ts: stdioトランスポート起動
- src/transports/streamable-http.ts: Streamable HTTPトランスポート起動
- src/transports/sse.ts: SSEトランスポート起動（非推奨）
- src/index.ts: エントリーポイント(317行→28行に削減)
- JSON-RPCエラーレスポンスのヘルパー関数を共通化
