+++ src/transports/sse.ts
@@ -1,0 +1,36 @@
+/**
+ * SSE トランスポートでの起動（非推奨）
+ */
+
+import express, { Request, Response } from 'express';
+import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse.js';
+import { createServer } from '../server.js';
+
+export async function startSSE(): Promise<void> {
+  const app = express();
+  const transports: { [sessionId: string]: SSEServerTransport } = {};
+
+  app.get('/sse', async (_: Request, res: Response) => {
+    const transport = new SSEServerTransport('/messages', res);
+    transports[transport.sessionId] = transport;
+    res.on('close', () => {
+      delete transports[transport.sessionId];
+    });
+    const sessionServer = createServer();
+    await sessionServer.connect(transport);
+  });
+
+  app.post('/messages', async (req: Request, res: Response) => {
+    const sessionId = req.query.sessionId as string;
+    const transport = transports[sessionId];
+    if (transport) {
+      await transport.handlePostMessage(req, res);
+    } else {
+      res.status(400).send('No transport found for sessionId');
+    }
+  });
+
+  const PORT = process.env.PORT || 3000;
+  app.listen(PORT);
+  console.error(`Redash MCP Server running on SSE at http://localhost:${PORT}`);
+}
