+++ src/transports/stdio.ts
@@ -1,0 +1,13 @@
+/**
+ * stdio トランスポートでの起動
+ */
+
+import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
+import { createServer } from '../server.js';
+
+export async function startStdio(): Promise<void> {
+  const server = createServer();
+  const transport = new StdioServerTransport();
+  await server.connect(transport);
+  console.error('Redash MCP Server running on stdio');
+}
